requirements.txt
tqdm
scipy
minio
flask
loguru
gevent
pillow
onnxsim
gunicorn
matplotlib
opencv-python
pycocotools
tritonclient[all]
pulsar-client==2.10.2
